163: Rename `rmoods` to `nlp`
